Extracted calculateEdges to function
