+++ src/Visualizer/helpers/index.ts
@@ -10,3 +10,4 @@ export * from "./setEdgeClassName";
 export * from "./logTablePositions";
 export * from "./loadDatabaseConfig";
 export * from "./loadDatabases";
+export * from "./calculateEdges";
